Add a function to get the filter name
The logic to get the name (if provided) to include in the search pattern
currently has 3 branches. While I don't expect there to be more, this
will be a lot of move to copy and paste as new search types get added to
this plugin. Centralizing the logic will simplify things down the road.