Add a command to search in all open buffers
This will add a new `:TODOBuffers` command. It works just like `:TODO`
except it will search across all open buffers, not just the current one.

To keep the function signatures and logic simple, `s:todo` is being
renamed to `s:todocurrent` and a new `s:todobuffers` is being added. In
addition to using a different command (`lvimgrepadd` instead of
`lvimgrep`), `s:todobuffers` clears the local list before adding to it.